+++ openbas-api/src/main/java/io/openbas/migration/V2_76__Add_reply_to_scenarios_and_exercises.java
@@ -7,12 +7,12 @@ import org.springframework.stereotype.Component;
 import java.sql.Statement;
 
 @Component
-public class V2_76__Add_reply_to_scenarios extends BaseJavaMigration {
+public class V2_76__Add_reply_to_scenarios_and_exercises extends BaseJavaMigration {
 
     @Override
     public void migrate(Context context) throws Exception {
         Statement select = context.getConnection().createStatement();
-        // Exercise
+        // Scenario
         select.execute("""
                 create table scenario_mails_reply_to
                 (
@@ -20,6 +20,16 @@ public class V2_76__Add_reply_to_scenarios extends BaseJavaMigration {
                         constraint fk_scenario_id
                             references scenarios,
                     scenario_reply_to varchar(255)
+                );
+                """);
+        // Exercise
+        select.execute("""
+                create table exercise_mails_reply_to
+                (
+                    exercise_id varchar(255) not null
+                        constraint fk_exercise_id
+                            references exercises,
+                    exercise_reply_to varchar(255)
                 );
                 """);
     }
